requirements.txt
sounddevice
numpy
openai-whisper
nltk
streamlit
SpeechRecognition
soundfile
